feature : 추천상품 호출 api 구현

+++ backend/Business/src/main/java/com/picky/business/recommend/domain/entity/Recommended.java
@@ -15,10 +15,10 @@ import javax.persistence.*;
 public class Recommended {
     @Id
     @Column(name = "user_id")
-    private int userId;
+    private Long userId;
 
     @Column(name = "product_id")
-    private int productId;
+    private Long productId;
 
     private float rating;
 }
